Removing Framework\Application
